Refactor the SSE decoding code to the SSE layer

+++ crates/utils/src/os/windows.rs
@@ -40,7 +40,7 @@ pub fn get_info(p: impl AsRef<Path>) -> std::io::Result<DiskInfo> {
             Some(&mut total_number_of_bytes),
             Some(&mut total_number_of_free_bytes),
         )
-        .map_err(|e| Error::from_raw_os_error(e.code().0 as i32))?;
+        .map_err(|e| Error::from_raw_os_error(e.code().0))?;
     }
 
     let total = total_number_of_bytes;
@@ -66,7 +66,7 @@ pub fn get_info(p: impl AsRef<Path>) -> std::io::Result<DiskInfo> {
             Some(&mut number_of_free_clusters),
             Some(&mut total_number_of_clusters),
         )
-        .map_err(|e| Error::from_raw_os_error(e.code().0 as i32))?;
+        .map_err(|e| Error::from_raw_os_error(e.code().0))?;
     }
 
     Ok(DiskInfo {
@@ -94,7 +94,7 @@ fn get_volume_name(v: &[u16]) -> std::io::Result<Vec<u16>> {
 
     unsafe {
         GetVolumePathNameW(windows::core::PCWSTR::from_raw(v.as_ptr()), &mut volume_name_buffer)
-            .map_err(|e| Error::from_raw_os_error(e.code().0 as i32))?;
+            .map_err(|e| Error::from_raw_os_error(e.code().0))?;
     }
 
     let len = volume_name_buffer
@@ -137,7 +137,7 @@ fn get_fs_type(p: &[u16]) -> std::io::Result<String> {
             Some(&mut file_system_flags),
             Some(&mut file_system_name_buffer),
         )
-        .map_err(|e| Error::from_raw_os_error(e.code().0 as i32))?;
+        .map_err(|e| Error::from_raw_os_error(e.code().0))?;
     }
 
     Ok(utf16_to_string(&file_system_name_buffer))
